+++ CHANGELOG.md
@@ -22,7 +22,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- Nothing.
+- [#1](https://github.com/laminas/laminas-json-server/pull/1) Allows optional `data` field in Error object to be  omitted
+
 
 ## 3.2.0 - 2019-10-17